+++ src/components/input/StyledInput.tsx
@@ -1,5 +1,4 @@
 import styled from 'styled-components';
-import { darken } from '../../scripts/color-functions';
 
 export default styled.input`
   padding: 0.4rem 0.6rem;
@@ -26,6 +25,6 @@ export default styled.input`
   }
   &:disabled {
     cursor: not-allowed;
-    color: ${({ theme }) => darken(theme.shade.darkest, 0.4)};
+    color: ${({ theme }) => theme.shade.darkest};
   }
 `;